requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
azure-monitor-opentelemetry 

azure-functions
confluent-kafka
confluent-kafka[schemaregistry]
orjson
jsonschema